requirements.txt
accelerate>=0.34.0,<1.0.0
bitsandbytes>=0.43.1; platform_system == "Linux"
datasets>=2.21.0,<3.0.0
openai>=1.40.0,<2.0.0
peft>=0.12.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
transformers>=4.44.0,<5.0.0
trl>=0.10.1,<1.0.0
